tests/test_measurement_engine.py
"""
Tests for the measurement engine module.
"""

import pytest
import numpy as np
from app.core.measurement_engine import MeasurementEngine, MeasurementProcessor

def test_measurement_engine_initialization():
    """Test initialization of MeasurementEngine."""
    engine = MeasurementEngine()
    assert engine.calibration_height is None
    assert engine.pose_detector is not None

def test_measurement_processor_initialization():
    """Test initialization of MeasurementProcessor."""
    processor = MeasurementProcessor()
    assert isinstance(processor.measurement_history, list)
    assert len(processor.measurement_history) == 0

def test_detect_pose():
    """Test pose detection functionality."""
    engine = MeasurementEngine()
    # Create a dummy image
    dummy_image = np.zeros((480, 640, 3), dtype=np.uint8)
    landmarks = engine.detect_pose(dummy_image)
    # TODO: Add proper assertions once implemented
    assert True

def test_extract_measurements():
    """Test measurement extraction from landmarks."""
    engine = MeasurementEngine()
    # TODO: Create dummy landmarks and test measurement extraction
    assert True

def test_measurement_validation():
    """Test measurement validation logic."""
    processor = MeasurementProcessor()
    # TODO: Add test cases for valid and invalid measurements
    assert True

def test_measurement_smoothing():
    """Test measurement smoothing functionality."""
    processor = MeasurementProcessor()
    # TODO: Add test cases for smoothing algorithm
    assert True

# Explanation:
# This test file provides basic test stubs for the measurement engine:
# 1. Tests for initialization of both main classes
# 2. Tests for core functionality (pose detection, measurement extraction)
# 3. Tests for data processing (validation, smoothing)
#
# Next steps:
# 1. Add proper test data and assertions
# 2. Implement edge case testing
# 3. Add performance benchmarks
# 4. Add integration tests 