requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
python-multipart==0.0.9
pydantic==2.6.1
starlette==0.36.3
typing-extensions==4.9.0
opencv-python==4.9.0.80
numpy==1.26.3
mediapipe==0.10.9
scikit-image==0.22.0
